# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hull Local API
  version: "1.0.0"
  description: |
    The contract between `hulld` and its clients (CLI, and the native GUIs from
    Phase 3 Wave 2 onward). This document is the FROZEN contract: every native
    client builds against it. Keep it in sync with `internal/api` , the
    coverage test `internal/api/openapi_test.go` fails if a registered route is
    missing here.

    Transport (ADR 0006): HTTP on `127.0.0.1:<port>` where `<port>` and a bearer
    token are read from `~/.hull/daemon.json` (written 0600 by a running
    daemon). Auth accepts `Authorization: Bearer <token>` or `?token=<token>`
    (the latter for EventSource/SSE, which cannot set headers). The server
    reflects any Origin for CORS , safe because the token, not the origin, is
    the access control and it only listens on loopback.

    Long operations (create, import, add-service, link, rebuild, reset) are
    JOBS: a POST returns a job reference; observe it by polling `GET
    /v1/jobs/{id}` or streaming `GET /v1/jobs/{id}/stream`.
servers:
  - url: http://127.0.0.1:{port}
    description: Local daemon (port from ~/.hull/daemon.json)
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
  - tokenQuery: []
...
paths:
...
  /v1/detect:
    get:
      tags: [projects]
      summary: File-based detection for an unmanaged folder under a root
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetectInfo" }
...
  /v1/projects/{name}:
    patch:
      tags: [projects]
      summary: Update a managed project's fields (php/domain/serve)
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchProjectRequest" }
      responses:
        "204": { description: Updated }
    delete:
      tags: [projects]
      summary: Destroy a project
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: files, in: query, required: false, schema: { type: boolean }, description: also delete the project directory }
      responses:
        "204": { description: Destroyed }
  /v1/projects/{name}/open:
    post:
      tags: [projects]
      summary: Open the project folder or in the configured editor
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OpenRequest" }
      responses:
        "204": { description: Opened }
  /v1/projects/{name}/group:
    post:
      tags: [groups]
      summary: Assign a project to a virtual group ("" to ungroup)
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { group: { type: string } }
      responses:
        "204": { description: Assigned }
...
  /v1/config:
    get:
      tags: [config]
      summary: Global configuration
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigInfo" }
    put:
      tags: [config]
      summary: Replace global configuration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigInfo" }
      responses:
        "200":
          description: Updated (with restart_required)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigInfo" }
  /v1/groups:
    get:
      tags: [groups]
      summary: The virtual-group store
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupStore" }
    put:
      tags: [groups]
      summary: Replace the virtual-group store
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GroupStore" }
      responses:
        "204": { description: Saved }
...
  /v1/logs:
    get:
      tags: [system]
      summary: Stream container logs (SSE)
      parameters:
        - { name: project, in: query, required: false, schema: { type: string } }
        - { name: service, in: query, required: false, schema: { type: string } }
        - { name: tail, in: query, required: false, schema: { type: integer } }
      responses:
        "200": { description: SSE stream, content: { text/event-stream: { schema: { type: string } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    tokenQuery:
      type: apiKey
      in: query
      name: token
  schemas:
...
    DetectInfo:
      type: object
      required: [kind, template, php_kind]
      properties:
        kind: { type: string }
        template: { type: string }
        php: { type: string }
        db: { type: string }
        database: { type: string }
        redis: { type: boolean }
        extras: { type: array, items: { type: string } }
        php_kind: { type: boolean }
...
    PatchProjectRequest:
      type: object
      properties:
        php: { type: string, nullable: true }
        domain: { type: string, nullable: true }
        serve: { type: boolean, nullable: true }
    OpenRequest:
      type: object
      required: [target]
      properties:
        target: { type: string, enum: [folder, editor] }
...
    ConfigInfo:
      type: object
      required: [tld, roots, defaults]
      properties:
        tld: { type: string }
        roots: { type: array, items: { type: string } }
        projects:
          type: array
          items: { type: string }
          description: "Individually-registered project directories (imported in place, outside any root). Absent in a PUT leaves them unchanged; an explicit list (including empty) replaces them."
        loopback: { type: string }
        defaults:
          type: object
          properties:
            php: { type: string }
            editor: { type: string }
            db_tool: { type: string }
        restart_required: { type: array, items: { type: string } }
...
    GroupStore:
      type: object
      description: "Virtual-group store keyed by absolute root path; opaque to clients beyond round-tripping."
      additionalProperties: true
...